fix(libsaferequest): make the branch squash-merge clean against #462/#463
The team merges with "Squash and merge", which drops the #462 -> #463 -> #464
ancestry. Git then falls back to origin/main as the merge base and sees #462's
changes -- already present here -- as competing additions, producing two
conflicts that a plain merge chain never hits:

  CONFLICT (content): internal/snapshot/transport/http.go
  CONFLICT (add/add): pkg/libsaferequest/client/http_test.go

Both were purely structural, so reshape them instead of changing behaviour:

  - http_test.go is created by #462 too, so an add/add conflicts unless both
    sides match byte for byte. Restore it to #462's exact content and move the
    7 SetTLSCAData tests and their 3 helpers to a new http_tls_test.go, which
    exists on neither side of main and merges as a clean single-side add.

  - The "keep in sync" note sat at the end of a doc comment block #462 had just
    added, so the two additions were adjacent and conflicted. Move it just
    inside SetTLSCAData, a region #462 does not touch.

No production code changes: pkg/libsaferequest/client/http.go is untouched and
internal/snapshot/transport/http.go now matches #463 apart from the relocated
comment. All 8 tests are preserved and still pass under -race.

Verified from origin/main: squash #462, squash #463, merge #464, merge #465 --
all four clean, then build, vet and go test -race clean over
./internal/snapshot/... and ./pkg/libsaferequest/...
